+++ packages/backend/src/middlewares/combinedAuth/index.ts
@@ -1,1 +1,0 @@
-export { combinedAuth } from "./combinedAuth";