+++ packages/frontend/src/components/LoginPage/index.ts
@@ -1,2 +1,1 @@
-export { useVkLogin } from "./LoginPage";
-
+export { useVkFloatingOneTap } from "./LoginPage";