+++ packages/frontend/src/types/global.d.ts
@@ -47,27 +47,6 @@ declare global {
       WidgetEvents: { ERROR: string };
       OneTapInternalEvents: { LOGIN_SUCCESS: string };
     };
-    Telegram?: {
-      WebApp: {
-        initData: string;
-        initDataUnsafe: {
-          query_id?: string;
-          user?: Record<string, unknown>;
-          receiver?: Record<string, unknown>;
-          chat?: {
-            id: number;
-            type: string;
-            title: string;
-          };
-          can_send_after?: number;
-          auth_date: number;
-          hash: string;
-        };
-        ready: () => void;
-        close: () => void;
-        expand: () => void;
-        themeParams: Record<string, string>;
-      };
-    };
+
   }
 }